+++ types/plugin.d.ts
@@ -1,9 +1,7 @@
-import { PluginClass } from 'ultimate-crosscode-typedefs/modloader/mod';
-import type { Mod1 } from 'ccmodmanager/src/types';
+import type { PluginClass } from 'ultimate-crosscode-typedefs/modloader/mod';
+import type { Mod1 } from 'ccmodmanager/types/types';
 declare global {
-    namespace sc {
-        var nwjsManager: CCNwjsManager;
-    }
+    var nwjsManager: CCNwjsManager;
 }
 export default class CCNwjsManager implements PluginClass {
     static dir: string;
@@ -16,4 +14,8 @@ export default class CCNwjsManager implements PluginClass {
     install(version: string, sdk: boolean): Promise<void>;
     private spawnLinuxScript;
     private spawnWindowsScript;
+    /** Grabs the SDK argument from the current installation.
+     * @param version {string} If this version < current NW.js version, install this NW.js version
+     * @returns {boolean} True if no installation is needed */
+    ensureVersion(version: string): boolean;
 }