+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 # Change Log
 
 ## [Unreleased]
+
+### Added
+
+- Added `nwjsManager.ensureVersion` function
+
 ## [1.0.2] 2024-06-13
 ## [1.0.1] 2024-06-12
 ## [1.0.0] 2024-06-10